requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
openai>=1.14.0
faiss-cpu>=1.8.0
tiktoken>=0.6.0
PyPDF2>=3.0.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.27
aiosqlite>=0.19.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.0.0
greenlet>=3.0.3
